Move Serialize helper functions into their own file.
